LayoutTests/imported/w3c/web-platform-tests/css/css-values/calc-size/animation/interpolate-size-min-width-interpolation-expected.txt


FAIL CSS Transitions: property <min-width> from [initial] to [20px] at (-0.3) should be [calc-size(auto, -6px + size * 1.3)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [initial] to [20px] at (0) should be [calc-size(auto, 0px + size * 1)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [initial] to [20px] at (0.3) should be [calc-size(auto, 6px + size * 0.7)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [initial] to [20px] at (0.6) should be [calc-size(auto, 12px + size * 0.4)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [initial] to [20px] at (1) should be [calc-size(auto, 20px + size * 0)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [initial] to [20px] at (1.5) should be [calc-size(auto, 30px + size * -0.5)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [initial] to [20px] at (-0.3) should be [calc-size(auto, -6px + size * 1.3)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [initial] to [20px] at (0) should be [calc-size(auto, 0px + size * 1)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [initial] to [20px] at (0.3) should be [calc-size(auto, 6px + size * 0.7)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [initial] to [20px] at (0.6) should be [calc-size(auto, 12px + size * 0.4)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [initial] to [20px] at (1) should be [calc-size(auto, 20px + size * 0)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [initial] to [20px] at (1.5) should be [calc-size(auto, 30px + size * -0.5)] assert_equals: expected "10px " but got "20px "
FAIL CSS Animations: property <min-width> from [initial] to [20px] at (-0.3) should be [calc-size(auto, -6px + size * 1.3)] assert_equals: expected "10px " but got "0px "
FAIL CSS Animations: property <min-width> from [initial] to [20px] at (0) should be [calc-size(auto, 0px + size * 1)] assert_equals: expected "10px " but got "0px "
FAIL CSS Animations: property <min-width> from [initial] to [20px] at (0.3) should be [calc-size(auto, 6px + size * 0.7)] assert_equals: expected "10px " but got "0px "
FAIL CSS Animations: property <min-width> from [initial] to [20px] at (0.6) should be [calc-size(auto, 12px + size * 0.4)] assert_equals: expected "10px " but got "20px "
FAIL CSS Animations: property <min-width> from [initial] to [20px] at (1) should be [calc-size(auto, 20px + size * 0)] assert_equals: expected "10px " but got "20px "
FAIL CSS Animations: property <min-width> from [initial] to [20px] at (1.5) should be [calc-size(auto, 30px + size * -0.5)] assert_equals: expected "10px " but got "20px "
FAIL Web Animations: property <min-width> from [initial] to [20px] at (-0.3) should be [calc-size(auto, -6px + size * 1.3)] assert_equals: expected "10px " but got "0px "
FAIL Web Animations: property <min-width> from [initial] to [20px] at (0) should be [calc-size(auto, 0px + size * 1)] assert_equals: expected "10px " but got "0px "
FAIL Web Animations: property <min-width> from [initial] to [20px] at (0.3) should be [calc-size(auto, 6px + size * 0.7)] assert_equals: expected "10px " but got "0px "
FAIL Web Animations: property <min-width> from [initial] to [20px] at (0.6) should be [calc-size(auto, 12px + size * 0.4)] assert_equals: expected "10px " but got "20px "
FAIL Web Animations: property <min-width> from [initial] to [20px] at (1) should be [calc-size(auto, 20px + size * 0)] assert_equals: expected "10px " but got "20px "
FAIL Web Animations: property <min-width> from [initial] to [20px] at (1.5) should be [calc-size(auto, 30px + size * -0.5)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [unset] to [20px] at (-0.3) should be [calc-size(auto, -6px + size * 1.3)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [unset] to [20px] at (0) should be [calc-size(auto, 0px + size * 1)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [unset] to [20px] at (0.3) should be [calc-size(auto, 6px + size * 0.7)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [unset] to [20px] at (0.6) should be [calc-size(auto, 12px + size * 0.4)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [unset] to [20px] at (1) should be [calc-size(auto, 20px + size * 0)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [unset] to [20px] at (1.5) should be [calc-size(auto, 30px + size * -0.5)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [unset] to [20px] at (-0.3) should be [calc-size(auto, -6px + size * 1.3)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [unset] to [20px] at (0) should be [calc-size(auto, 0px + size * 1)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [unset] to [20px] at (0.3) should be [calc-size(auto, 6px + size * 0.7)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [unset] to [20px] at (0.6) should be [calc-size(auto, 12px + size * 0.4)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [unset] to [20px] at (1) should be [calc-size(auto, 20px + size * 0)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [unset] to [20px] at (1.5) should be [calc-size(auto, 30px + size * -0.5)] assert_equals: expected "10px " but got "20px "
FAIL CSS Animations: property <min-width> from [unset] to [20px] at (-0.3) should be [calc-size(auto, -6px + size * 1.3)] assert_equals: expected "10px " but got "0px "
FAIL CSS Animations: property <min-width> from [unset] to [20px] at (0) should be [calc-size(auto, 0px + size * 1)] assert_equals: expected "10px " but got "0px "
FAIL CSS Animations: property <min-width> from [unset] to [20px] at (0.3) should be [calc-size(auto, 6px + size * 0.7)] assert_equals: expected "10px " but got "0px "
FAIL CSS Animations: property <min-width> from [unset] to [20px] at (0.6) should be [calc-size(auto, 12px + size * 0.4)] assert_equals: expected "10px " but got "20px "
FAIL CSS Animations: property <min-width> from [unset] to [20px] at (1) should be [calc-size(auto, 20px + size * 0)] assert_equals: expected "10px " but got "20px "
FAIL CSS Animations: property <min-width> from [unset] to [20px] at (1.5) should be [calc-size(auto, 30px + size * -0.5)] assert_equals: expected "10px " but got "20px "
FAIL Web Animations: property <min-width> from [unset] to [20px] at (-0.3) should be [calc-size(auto, -6px + size * 1.3)] assert_equals: expected "10px " but got "0px "
FAIL Web Animations: property <min-width> from [unset] to [20px] at (0) should be [calc-size(auto, 0px + size * 1)] assert_equals: expected "10px " but got "0px "
FAIL Web Animations: property <min-width> from [unset] to [20px] at (0.3) should be [calc-size(auto, 6px + size * 0.7)] assert_equals: expected "10px " but got "0px "
FAIL Web Animations: property <min-width> from [unset] to [20px] at (0.6) should be [calc-size(auto, 12px + size * 0.4)] assert_equals: expected "10px " but got "20px "
FAIL Web Animations: property <min-width> from [unset] to [20px] at (1) should be [calc-size(auto, 20px + size * 0)] assert_equals: expected "10px " but got "20px "
FAIL Web Animations: property <min-width> from [unset] to [20px] at (1.5) should be [calc-size(auto, 30px + size * -0.5)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [auto] to [20px] at (-0.3) should be [calc-size(auto, -6px + size * 1.3)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [auto] to [20px] at (0) should be [calc-size(auto, 0px + size * 1)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [auto] to [20px] at (0.3) should be [calc-size(auto, 6px + size * 0.7)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [auto] to [20px] at (0.6) should be [calc-size(auto, 12px + size * 0.4)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [auto] to [20px] at (1) should be [calc-size(auto, 20px + size * 0)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [auto] to [20px] at (1.5) should be [calc-size(auto, 30px + size * -0.5)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [auto] to [20px] at (-0.3) should be [calc-size(auto, -6px + size * 1.3)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [auto] to [20px] at (0) should be [calc-size(auto, 0px + size * 1)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [auto] to [20px] at (0.3) should be [calc-size(auto, 6px + size * 0.7)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [auto] to [20px] at (0.6) should be [calc-size(auto, 12px + size * 0.4)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [auto] to [20px] at (1) should be [calc-size(auto, 20px + size * 0)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions with transition: all: property <min-width> from [auto] to [20px] at (1.5) should be [calc-size(auto, 30px + size * -0.5)] assert_equals: expected "10px " but got "20px "
FAIL CSS Animations: property <min-width> from [auto] to [20px] at (-0.3) should be [calc-size(auto, -6px + size * 1.3)] assert_equals: expected "10px " but got "0px "
FAIL CSS Animations: property <min-width> from [auto] to [20px] at (0) should be [calc-size(auto, 0px + size * 1)] assert_equals: expected "10px " but got "0px "
FAIL CSS Animations: property <min-width> from [auto] to [20px] at (0.3) should be [calc-size(auto, 6px + size * 0.7)] assert_equals: expected "10px " but got "0px "
FAIL CSS Animations: property <min-width> from [auto] to [20px] at (0.6) should be [calc-size(auto, 12px + size * 0.4)] assert_equals: expected "10px " but got "20px "
FAIL CSS Animations: property <min-width> from [auto] to [20px] at (1) should be [calc-size(auto, 20px + size * 0)] assert_equals: expected "10px " but got "20px "
FAIL CSS Animations: property <min-width> from [auto] to [20px] at (1.5) should be [calc-size(auto, 30px + size * -0.5)] assert_equals: expected "10px " but got "20px "
FAIL Web Animations: property <min-width> from [auto] to [20px] at (-0.3) should be [calc-size(auto, -6px + size * 1.3)] assert_equals: expected "10px " but got "0px "
FAIL Web Animations: property <min-width> from [auto] to [20px] at (0) should be [calc-size(auto, 0px + size * 1)] assert_equals: expected "10px " but got "0px "
FAIL Web Animations: property <min-width> from [auto] to [20px] at (0.3) should be [calc-size(auto, 6px + size * 0.7)] assert_equals: expected "10px " but got "0px "
FAIL Web Animations: property <min-width> from [auto] to [20px] at (0.6) should be [calc-size(auto, 12px + size * 0.4)] assert_equals: expected "10px " but got "20px "
FAIL Web Animations: property <min-width> from [auto] to [20px] at (1) should be [calc-size(auto, 20px + size * 0)] assert_equals: expected "10px " but got "20px "
FAIL Web Animations: property <min-width> from [auto] to [20px] at (1.5) should be [calc-size(auto, 30px + size * -0.5)] assert_equals: expected "10px " but got "20px "
FAIL CSS Transitions: property <min-width> from [0px] to [stretch] at (-0.3) should be [calc-size(stretch, 0px + size * -0.3)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <min-width> from [0px] to [stretch] at (0) should be [calc-size(stretch, 0px + size * 0)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <min-width> from [0px] to [stretch] at (0.3) should be [calc-size(stretch, 0px + size * 0.3)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <min-width> from [0px] to [stretch] at (0.6) should be [calc-size(stretch, 0px + size * 0.6)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <min-width> from [0px] to [stretch] at (1) should be [calc-size(stretch, 0px + size * 1)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <min-width> from [0px] to [stretch] at (1.5) should be [calc-size(stretch, 0px + size * 1.5)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (-0.3) should be [calc-size(stretch, 0px + size * -0.3)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (0) should be [calc-size(stretch, 0px + size * 0)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (0.3) should be [calc-size(stretch, 0px + size * 0.3)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (0.6) should be [calc-size(stretch, 0px + size * 0.6)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (1) should be [calc-size(stretch, 0px + size * 1)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (1.5) should be [calc-size(stretch, 0px + size * 1.5)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <min-width> from [0px] to [stretch] at (-0.3) should be [calc-size(stretch, 0px + size * -0.3)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <min-width> from [0px] to [stretch] at (0) should be [calc-size(stretch, 0px + size * 0)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <min-width> from [0px] to [stretch] at (0.3) should be [calc-size(stretch, 0px + size * 0.3)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <min-width> from [0px] to [stretch] at (0.6) should be [calc-size(stretch, 0px + size * 0.6)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <min-width> from [0px] to [stretch] at (1) should be [calc-size(stretch, 0px + size * 1)] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <min-width> from [0px] to [stretch] at (1.5) should be [calc-size(stretch, 0px + size * 1.5)] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <min-width> from [0px] to [stretch] at (-0.3) should be [calc-size(stretch, 0px + size * -0.3)] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <min-width> from [0px] to [stretch] at (0) should be [calc-size(stretch, 0px + size * 0)] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <min-width> from [0px] to [stretch] at (0.3) should be [calc-size(stretch, 0px + size * 0.3)] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <min-width> from [0px] to [stretch] at (0.6) should be [calc-size(stretch, 0px + size * 0.6)] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <min-width> from [0px] to [stretch] at (1) should be [calc-size(stretch, 0px + size * 1)] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <min-width> from [0px] to [stretch] at (1.5) should be [calc-size(stretch, 0px + size * 1.5)] assert_true: 'to' value should be supported expected true got false
PASS CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [min-content] to [fit-content] at (-0.3) should be [min-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [min-content] to [fit-content] at (0) should be [min-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [min-content] to [fit-content] at (0.3) should be [min-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [min-content] to [fit-content] at (0.5) should be [fit-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [min-content] to [fit-content] at (0.6) should be [fit-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [min-content] to [fit-content] at (1) should be [fit-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [min-content] to [fit-content] at (1.5) should be [fit-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [min-content] to [fit-content] at (-0.3) should be [min-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [min-content] to [fit-content] at (0) should be [min-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [min-content] to [fit-content] at (0.3) should be [min-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [min-content] to [fit-content] at (0.5) should be [fit-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [min-content] to [fit-content] at (0.6) should be [fit-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [min-content] to [fit-content] at (1) should be [fit-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [min-content] to [fit-content] at (1.5) should be [fit-content]
PASS CSS Transitions: property <min-width> from [min-content] to [fit-content] at (-0.3) should be [fit-content]
PASS CSS Transitions: property <min-width> from [min-content] to [fit-content] at (0) should be [fit-content]
PASS CSS Transitions: property <min-width> from [min-content] to [fit-content] at (0.3) should be [fit-content]
PASS CSS Transitions: property <min-width> from [min-content] to [fit-content] at (0.5) should be [fit-content]
PASS CSS Transitions: property <min-width> from [min-content] to [fit-content] at (0.6) should be [fit-content]
PASS CSS Transitions: property <min-width> from [min-content] to [fit-content] at (1) should be [fit-content]
PASS CSS Transitions: property <min-width> from [min-content] to [fit-content] at (1.5) should be [fit-content]
PASS CSS Transitions with transition: all: property <min-width> from [min-content] to [fit-content] at (-0.3) should be [fit-content]
PASS CSS Transitions with transition: all: property <min-width> from [min-content] to [fit-content] at (0) should be [fit-content]
PASS CSS Transitions with transition: all: property <min-width> from [min-content] to [fit-content] at (0.3) should be [fit-content]
PASS CSS Transitions with transition: all: property <min-width> from [min-content] to [fit-content] at (0.5) should be [fit-content]
PASS CSS Transitions with transition: all: property <min-width> from [min-content] to [fit-content] at (0.6) should be [fit-content]
PASS CSS Transitions with transition: all: property <min-width> from [min-content] to [fit-content] at (1) should be [fit-content]
PASS CSS Transitions with transition: all: property <min-width> from [min-content] to [fit-content] at (1.5) should be [fit-content]
PASS CSS Animations: property <min-width> from [min-content] to [fit-content] at (-0.3) should be [min-content]
PASS CSS Animations: property <min-width> from [min-content] to [fit-content] at (0) should be [min-content]
PASS CSS Animations: property <min-width> from [min-content] to [fit-content] at (0.3) should be [min-content]
PASS CSS Animations: property <min-width> from [min-content] to [fit-content] at (0.5) should be [fit-content]
PASS CSS Animations: property <min-width> from [min-content] to [fit-content] at (0.6) should be [fit-content]
PASS CSS Animations: property <min-width> from [min-content] to [fit-content] at (1) should be [fit-content]
PASS CSS Animations: property <min-width> from [min-content] to [fit-content] at (1.5) should be [fit-content]
PASS Web Animations: property <min-width> from [min-content] to [fit-content] at (-0.3) should be [min-content]
PASS Web Animations: property <min-width> from [min-content] to [fit-content] at (0) should be [min-content]
PASS Web Animations: property <min-width> from [min-content] to [fit-content] at (0.3) should be [min-content]
PASS Web Animations: property <min-width> from [min-content] to [fit-content] at (0.5) should be [fit-content]
PASS Web Animations: property <min-width> from [min-content] to [fit-content] at (0.6) should be [fit-content]
PASS Web Animations: property <min-width> from [min-content] to [fit-content] at (1) should be [fit-content]
PASS Web Animations: property <min-width> from [min-content] to [fit-content] at (1.5) should be [fit-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [auto] to [max-content] at (-0.3) should be [auto]
PASS CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [auto] to [max-content] at (0) should be [auto]
PASS CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [auto] to [max-content] at (0.3) should be [auto]
PASS CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [auto] to [max-content] at (0.5) should be [max-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [auto] to [max-content] at (0.6) should be [max-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [auto] to [max-content] at (1) should be [max-content]
PASS CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [auto] to [max-content] at (1.5) should be [max-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [auto] to [max-content] at (-0.3) should be [auto]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [auto] to [max-content] at (0) should be [auto]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [auto] to [max-content] at (0.3) should be [auto]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [auto] to [max-content] at (0.5) should be [max-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [auto] to [max-content] at (0.6) should be [max-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [auto] to [max-content] at (1) should be [max-content]
PASS CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [auto] to [max-content] at (1.5) should be [max-content]
PASS CSS Transitions: property <min-width> from [auto] to [max-content] at (-0.3) should be [max-content]
PASS CSS Transitions: property <min-width> from [auto] to [max-content] at (0) should be [max-content]
PASS CSS Transitions: property <min-width> from [auto] to [max-content] at (0.3) should be [max-content]
PASS CSS Transitions: property <min-width> from [auto] to [max-content] at (0.5) should be [max-content]
PASS CSS Transitions: property <min-width> from [auto] to [max-content] at (0.6) should be [max-content]
PASS CSS Transitions: property <min-width> from [auto] to [max-content] at (1) should be [max-content]
PASS CSS Transitions: property <min-width> from [auto] to [max-content] at (1.5) should be [max-content]
PASS CSS Transitions with transition: all: property <min-width> from [auto] to [max-content] at (-0.3) should be [max-content]
PASS CSS Transitions with transition: all: property <min-width> from [auto] to [max-content] at (0) should be [max-content]
PASS CSS Transitions with transition: all: property <min-width> from [auto] to [max-content] at (0.3) should be [max-content]
PASS CSS Transitions with transition: all: property <min-width> from [auto] to [max-content] at (0.5) should be [max-content]
PASS CSS Transitions with transition: all: property <min-width> from [auto] to [max-content] at (0.6) should be [max-content]
PASS CSS Transitions with transition: all: property <min-width> from [auto] to [max-content] at (1) should be [max-content]
PASS CSS Transitions with transition: all: property <min-width> from [auto] to [max-content] at (1.5) should be [max-content]
PASS CSS Animations: property <min-width> from [auto] to [max-content] at (-0.3) should be [auto]
PASS CSS Animations: property <min-width> from [auto] to [max-content] at (0) should be [auto]
PASS CSS Animations: property <min-width> from [auto] to [max-content] at (0.3) should be [auto]
PASS CSS Animations: property <min-width> from [auto] to [max-content] at (0.5) should be [max-content]
PASS CSS Animations: property <min-width> from [auto] to [max-content] at (0.6) should be [max-content]
PASS CSS Animations: property <min-width> from [auto] to [max-content] at (1) should be [max-content]
PASS CSS Animations: property <min-width> from [auto] to [max-content] at (1.5) should be [max-content]
PASS Web Animations: property <min-width> from [auto] to [max-content] at (-0.3) should be [auto]
PASS Web Animations: property <min-width> from [auto] to [max-content] at (0) should be [auto]
PASS Web Animations: property <min-width> from [auto] to [max-content] at (0.3) should be [auto]
PASS Web Animations: property <min-width> from [auto] to [max-content] at (0.5) should be [max-content]
PASS Web Animations: property <min-width> from [auto] to [max-content] at (0.6) should be [max-content]
PASS Web Animations: property <min-width> from [auto] to [max-content] at (1) should be [max-content]
PASS Web Animations: property <min-width> from [auto] to [max-content] at (1.5) should be [max-content]
PASS CSS Transitions: property <min-width> from [fit-content] to neutral at (-0.3) should be [calc-size(fit-content, -3px + size * 1.3)]
PASS CSS Transitions: property <min-width> from [fit-content] to neutral at (0) should be [calc-size(fit-content, 0px + size * 1)]
PASS CSS Transitions: property <min-width> from [fit-content] to neutral at (0.3) should be [calc-size(fit-content, 3px + size * 0.7)]
PASS CSS Transitions: property <min-width> from [fit-content] to neutral at (0.6) should be [calc-size(fit-content, 6px + size * 0.4)]
PASS CSS Transitions: property <min-width> from [fit-content] to neutral at (1) should be [calc-size(fit-content, 10px + size * 0)]
PASS CSS Transitions: property <min-width> from [fit-content] to neutral at (1.5) should be [calc-size(fit-content, 15px + size * -0.5)]
PASS CSS Transitions with transition: all: property <min-width> from [fit-content] to neutral at (-0.3) should be [calc-size(fit-content, -3px + size * 1.3)]
PASS CSS Transitions with transition: all: property <min-width> from [fit-content] to neutral at (0) should be [calc-size(fit-content, 0px + size * 1)]
PASS CSS Transitions with transition: all: property <min-width> from [fit-content] to neutral at (0.3) should be [calc-size(fit-content, 3px + size * 0.7)]
PASS CSS Transitions with transition: all: property <min-width> from [fit-content] to neutral at (0.6) should be [calc-size(fit-content, 6px + size * 0.4)]
PASS CSS Transitions with transition: all: property <min-width> from [fit-content] to neutral at (1) should be [calc-size(fit-content, 10px + size * 0)]
PASS CSS Transitions with transition: all: property <min-width> from [fit-content] to neutral at (1.5) should be [calc-size(fit-content, 15px + size * -0.5)]
FAIL CSS Animations: property <min-width> from [fit-content] to neutral at (-0.3) should be [calc-size(fit-content, -3px + size * 1.3)] assert_equals: expected "10px " but got "fit - content "
FAIL CSS Animations: property <min-width> from [fit-content] to neutral at (0) should be [calc-size(fit-content, 0px + size * 1)] assert_equals: expected "10px " but got "fit - content "
FAIL CSS Animations: property <min-width> from [fit-content] to neutral at (0.3) should be [calc-size(fit-content, 3px + size * 0.7)] assert_equals: expected "10px " but got "fit - content "
PASS CSS Animations: property <min-width> from [fit-content] to neutral at (0.6) should be [calc-size(fit-content, 6px + size * 0.4)]
PASS CSS Animations: property <min-width> from [fit-content] to neutral at (1) should be [calc-size(fit-content, 10px + size * 0)]
PASS CSS Animations: property <min-width> from [fit-content] to neutral at (1.5) should be [calc-size(fit-content, 15px + size * -0.5)]
FAIL Web Animations: property <min-width> from [fit-content] to neutral at (-0.3) should be [calc-size(fit-content, -3px + size * 1.3)] assert_equals: expected "10px " but got "fit - content "
FAIL Web Animations: property <min-width> from [fit-content] to neutral at (0) should be [calc-size(fit-content, 0px + size * 1)] assert_equals: expected "10px " but got "fit - content "
FAIL Web Animations: property <min-width> from [fit-content] to neutral at (0.3) should be [calc-size(fit-content, 3px + size * 0.7)] assert_equals: expected "10px " but got "fit - content "
PASS Web Animations: property <min-width> from [fit-content] to neutral at (0.6) should be [calc-size(fit-content, 6px + size * 0.4)]
PASS Web Animations: property <min-width> from [fit-content] to neutral at (1) should be [calc-size(fit-content, 10px + size * 0)]
PASS Web Animations: property <min-width> from [fit-content] to neutral at (1.5) should be [calc-size(fit-content, 15px + size * -0.5)]
FAIL CSS Transitions: property <min-width> from neutral to [max-content] at (-0.3) should be [calc-size(max-content, 13px + size * -0.3)] assert_equals: expected "10px " but got "max - content "
FAIL CSS Transitions: property <min-width> from neutral to [max-content] at (0) should be [calc-size(max-content, 10px + size * 0)] assert_equals: expected "10px " but got "max - content "
FAIL CSS Transitions: property <min-width> from neutral to [max-content] at (0.3) should be [calc-size(max-content, 7px + size * 0.3)] assert_equals: expected "10px " but got "max - content "
FAIL CSS Transitions: property <min-width> from neutral to [max-content] at (0.6) should be [calc-size(max-content, 4px + size * 0.6)] assert_equals: expected "10px " but got "max - content "
FAIL CSS Transitions: property <min-width> from neutral to [max-content] at (1) should be [calc-size(max-content, 0px + size * 1)] assert_equals: expected "10px " but got "max - content "
FAIL CSS Transitions: property <min-width> from neutral to [max-content] at (1.5) should be [calc-size(max-content, -5px + size * 1.5)] assert_equals: expected "10px " but got "max - content "
FAIL CSS Transitions with transition: all: property <min-width> from neutral to [max-content] at (-0.3) should be [calc-size(max-content, 13px + size * -0.3)] assert_equals: expected "10px " but got "max - content "
FAIL CSS Transitions with transition: all: property <min-width> from neutral to [max-content] at (0) should be [calc-size(max-content, 10px + size * 0)] assert_equals: expected "10px " but got "max - content "
FAIL CSS Transitions with transition: all: property <min-width> from neutral to [max-content] at (0.3) should be [calc-size(max-content, 7px + size * 0.3)] assert_equals: expected "10px " but got "max - content "
FAIL CSS Transitions with transition: all: property <min-width> from neutral to [max-content] at (0.6) should be [calc-size(max-content, 4px + size * 0.6)] assert_equals: expected "10px " but got "max - content "
FAIL CSS Transitions with transition: all: property <min-width> from neutral to [max-content] at (1) should be [calc-size(max-content, 0px + size * 1)] assert_equals: expected "10px " but got "max - content "
FAIL CSS Transitions with transition: all: property <min-width> from neutral to [max-content] at (1.5) should be [calc-size(max-content, -5px + size * 1.5)] assert_equals: expected "10px " but got "max - content "
PASS CSS Animations: property <min-width> from neutral to [max-content] at (-0.3) should be [calc-size(max-content, 13px + size * -0.3)]
PASS CSS Animations: property <min-width> from neutral to [max-content] at (0) should be [calc-size(max-content, 10px + size * 0)]
PASS CSS Animations: property <min-width> from neutral to [max-content] at (0.3) should be [calc-size(max-content, 7px + size * 0.3)]
FAIL CSS Animations: property <min-width> from neutral to [max-content] at (0.6) should be [calc-size(max-content, 4px + size * 0.6)] assert_equals: expected "10px " but got "max - content "
FAIL CSS Animations: property <min-width> from neutral to [max-content] at (1) should be [calc-size(max-content, 0px + size * 1)] assert_equals: expected "10px " but got "max - content "
FAIL CSS Animations: property <min-width> from neutral to [max-content] at (1.5) should be [calc-size(max-content, -5px + size * 1.5)] assert_equals: expected "10px " but got "max - content "
PASS Web Animations: property <min-width> from neutral to [max-content] at (-0.3) should be [calc-size(max-content, 13px + size * -0.3)]
PASS Web Animations: property <min-width> from neutral to [max-content] at (0) should be [calc-size(max-content, 10px + size * 0)]
PASS Web Animations: property <min-width> from neutral to [max-content] at (0.3) should be [calc-size(max-content, 7px + size * 0.3)]
FAIL Web Animations: property <min-width> from neutral to [max-content] at (0.6) should be [calc-size(max-content, 4px + size * 0.6)] assert_equals: expected "10px " but got "max - content "
FAIL Web Animations: property <min-width> from neutral to [max-content] at (1) should be [calc-size(max-content, 0px + size * 1)] assert_equals: expected "10px " but got "max - content "
FAIL Web Animations: property <min-width> from neutral to [max-content] at (1.5) should be [calc-size(max-content, -5px + size * 1.5)] assert_equals: expected "10px " but got "max - content "
